added post security and authentication

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,17 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index','show']);
+    }
+
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -77,6 +88,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post=Post::find($id);
+
+        //check for correct user
+        if(auth()->user()->id !== $post->user->id){
+            return redirect('/posts')->with('error','Unauthorized request');
+        }
+
         return view('posts.edit')->with('post',$post);
     }
 
@@ -112,6 +129,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+         //check for correct user
+         if(auth()->user()->id !== $post->user->id){
+            return redirect('/posts')->with('error','Unauthorized request');
+        }
         $post->delete();
 
         return redirect('/posts')->with('success','Post Deleted');
